Move lock_check to separate file
This avoids:
- unused function lock_check warning
- duplicated code
